Switch location of load_bakery function
